requirements.txt
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2
rich==13.9.4
